// gen/ts/src/nama/api/v1/provider_pb.d.ts
export declare type ProviderConnectionTest = Message<"nama.api.v1.ProviderConnectionTest"> & {
  /**
   * @generated from field: nama.api.v1.ProviderConnectionStatus status = 1;
   */
  status: ProviderConnectionStatus;

  /**
   * @generated from field: string summary = 2;
   */
  summary: string;

  /**
   * @generated from field: optional string remote_name = 3;
   */
  remoteName?: string | undefined;

  /**
   * @generated from field: optional string remote_version = 4;
   */
  remoteVersion?: string | undefined;

  /**
   * @generated from field: repeated nama.api.v1.ProviderCapability capabilities = 5;
   */
  capabilities: ProviderCapability[];
};

/**
 * Describes the message nama.api.v1.ProviderConnectionTest.
 * Use `create(ProviderConnectionTestSchema)` to create a new message.
 */
export declare const ProviderConnectionTestSchema: GenMessage<ProviderConnectionTest>;

/**
 * @generated from message nama.api.v1.ListProviderTypesRequest
 */
export declare type ListProviderTypesRequest = Message<"nama.api.v1.ListProviderTypesRequest"> & {
  /**
   * @generated from field: uint32 page_size = 1;
   */
  pageSize: number;

  /**
   * @generated from field: string page_token = 2;
   */
  pageToken: string;
};

/**
 * Describes the message nama.api.v1.ListProviderTypesRequest.
 * Use `create(ListProviderTypesRequestSchema)` to create a new message.
 */
export declare const ListProviderTypesRequestSchema: GenMessage<ListProviderTypesRequest>;

/**
 * @generated from message nama.api.v1.ListProviderTypesResponse
 */
export declare type ListProviderTypesResponse = Message<"nama.api.v1.ListProviderTypesResponse"> & {
  /**
   * @generated from field: repeated nama.api.v1.ProviderType provider_types = 1;
   */
  providerTypes: ProviderType[];

  /**
   * @generated from field: string next_page_token = 2;
   */
  nextPageToken: string;
};

/**
 * Describes the message nama.api.v1.ListProviderTypesResponse.
 * Use `create(ListProviderTypesResponseSchema)` to create a new message.
 */
export declare const ListProviderTypesResponseSchema: GenMessage<ListProviderTypesResponse>;

/**
 * @generated from message nama.api.v1.ListProviderInstancesRequest
 */
export declare type ListProviderInstancesRequest = Message<"nama.api.v1.ListProviderInstancesRequest"> & {
  /**
   * @generated from field: uint32 page_size = 1;
   */
  pageSize: number;

  /**
   * @generated from field: string page_token = 2;
   */
  pageToken: string;
};

/**
 * Describes the message nama.api.v1.ListProviderInstancesRequest.
 * Use `create(ListProviderInstancesRequestSchema)` to create a new message.
 */
export declare const ListProviderInstancesRequestSchema: GenMessage<ListProviderInstancesRequest>;

/**
 * @generated from message nama.api.v1.ListProviderInstancesResponse
 */
export declare type ListProviderInstancesResponse = Message<"nama.api.v1.ListProviderInstancesResponse"> & {
  /**
   * @generated from field: repeated nama.api.v1.ProviderInstance provider_instances = 1;
   */
  providerInstances: ProviderInstance[];

  /**
   * @generated from field: string next_page_token = 2;
   */
  nextPageToken: string;
};

/**
 * Describes the message nama.api.v1.ListProviderInstancesResponse.
 * Use `create(ListProviderInstancesResponseSchema)` to create a new message.
 */
export declare const ListProviderInstancesResponseSchema: GenMessage<ListProviderInstancesResponse>;

/**
 * @generated from message nama.api.v1.GetProviderInstanceRequest
 */
export declare type GetProviderInstanceRequest = Message<"nama.api.v1.GetProviderInstanceRequest"> & {
  /**
   * @generated from field: string provider_instance_id = 1;
   */
  providerInstanceId: string;
};

/**
 * Describes the message nama.api.v1.GetProviderInstanceRequest.
 * Use `create(GetProviderInstanceRequestSchema)` to create a new message.
 */
export declare const GetProviderInstanceRequestSchema: GenMessage<GetProviderInstanceRequest>;

/**
 * @generated from message nama.api.v1.GetProviderInstanceResponse
 */
export declare type GetProviderInstanceResponse = Message<"nama.api.v1.GetProviderInstanceResponse"> & {
  /**
   * @generated from field: nama.api.v1.ProviderInstance provider_instance = 1;
   */
  providerInstance?: ProviderInstance | undefined;
};

/**
 * Describes the message nama.api.v1.GetProviderInstanceResponse.
 * Use `create(GetProviderInstanceResponseSchema)` to create a new message.
 */
export declare const GetProviderInstanceResponseSchema: GenMessage<GetProviderInstanceResponse>;

/**
 * @generated from message nama.api.v1.TestProviderConfigurationRequest
 */
export declare type TestProviderConfigurationRequest = Message<"nama.api.v1.TestProviderConfigurationRequest"> & {
  /**
   * @generated from field: string provider_type_id = 1;
   */
  providerTypeId: string;

  /**
   * @generated from field: google.protobuf.Struct configuration = 2;
   */
  configuration?: JsonObject | undefined;
};

/**
 * Describes the message nama.api.v1.TestProviderConfigurationRequest.
 * Use `create(TestProviderConfigurationRequestSchema)` to create a new message.
 */
export declare const TestProviderConfigurationRequestSchema: GenMessage<TestProviderConfigurationRequest>;

/**
 * @generated from message nama.api.v1.TestProviderConfigurationResponse
 */
export declare type TestProviderConfigurationResponse = Message<"nama.api.v1.TestProviderConfigurationResponse"> & {
  /**
   * @generated from field: nama.api.v1.ProviderConnectionTest result = 1;
   */
  result?: ProviderConnectionTest | undefined;
};

/**
 * Describes the message nama.api.v1.TestProviderConfigurationResponse.
 * Use `create(TestProviderConfigurationResponseSchema)` to create a new message.
 */
export declare const TestProviderConfigurationResponseSchema: GenMessage<TestProviderConfigurationResponse>;

/**
 * @generated from message nama.api.v1.TestProviderInstanceRequest
 */
export declare type TestProviderInstanceRequest = Message<"nama.api.v1.TestProviderInstanceRequest"> & {
  /**
   * @generated from field: string provider_instance_id = 1;
   */
  providerInstanceId: string;
};

/**
 * Describes the message nama.api.v1.TestProviderInstanceRequest.
 * Use `create(TestProviderInstanceRequestSchema)` to create a new message.
 */
export declare const TestProviderInstanceRequestSchema: GenMessage<TestProviderInstanceRequest>;

/**
 * @generated from message nama.api.v1.TestProviderInstanceResponse
 */
export declare type TestProviderInstanceResponse = Message<"nama.api.v1.TestProviderInstanceResponse"> & {
  /**
   * @generated from field: nama.api.v1.ProviderConnectionTest result = 1;
   */
  result?: ProviderConnectionTest | undefined;
};

/**
 * Describes the message nama.api.v1.TestProviderInstanceResponse.
 * Use `create(TestProviderInstanceResponseSchema)` to create a new message.
 */
export declare const TestProviderInstanceResponseSchema: GenMessage<TestProviderInstanceResponse>;

/**
 * @generated from message nama.api.v1.CreateProviderInstanceRequest
 */
export declare type CreateProviderInstanceRequest = Message<"nama.api.v1.CreateProviderInstanceRequest"> & {
  /**
   * @generated from field: string operation_id = 1;
   */
  operationId: string;

  /**
   * @generated from field: string provider_type_id = 2;
   */
  providerTypeId: string;

  /**
   * @generated from field: string display_name = 3;
   */
  displayName: string;

  /**
   * @generated from field: google.protobuf.Struct configuration = 4;
   */
  configuration?: JsonObject | undefined;

  /**
   * @generated from field: bool enabled = 5;
   */
  enabled: boolean;

  /**
   * @generated from field: optional uint32 sync_priority = 6;
   */
  syncPriority?: number | undefined;
};

/**
 * Describes the message nama.api.v1.CreateProviderInstanceRequest.
 * Use `create(CreateProviderInstanceRequestSchema)` to create a new message.
 */
export declare const CreateProviderInstanceRequestSchema: GenMessage<CreateProviderInstanceRequest>;

/**
 * @generated from message nama.api.v1.CreateProviderInstanceResponse
 */
export declare type CreateProviderInstanceResponse = Message<"nama.api.v1.CreateProviderInstanceResponse"> & {
  /**
   * @generated from field: nama.api.v1.ProviderInstance provider_instance = 1;
   */
  providerInstance?: ProviderInstance | undefined;
};

/**
 * Describes the message nama.api.v1.CreateProviderInstanceResponse.
 * Use `create(CreateProviderInstanceResponseSchema)` to create a new message.
 */
export declare const CreateProviderInstanceResponseSchema: GenMessage<CreateProviderInstanceResponse>;

/**
 * @generated from message nama.api.v1.UpdateProviderInstanceRequest
 */
export declare type UpdateProviderInstanceRequest = Message<"nama.api.v1.UpdateProviderInstanceRequest"> & {
  /**
   * @generated from field: string operation_id = 1;
   */
  operationId: string;

  /**
   * @generated from field: string provider_instance_id = 2;
   */
  providerInstanceId: string;

  /**
   * @generated from field: string expected_revision = 3;
   */
  expectedRevision: string;

  /**
   * @generated from field: optional string display_name = 4;
   */
  displayName?: string | undefined;

  /**
   * @generated from field: optional bool enabled = 5;
   */
  enabled?: boolean | undefined;

  /**
   * @generated from field: optional uint32 sync_priority = 6;
   */
  syncPriority?: number | undefined;

  /**
   * @generated from field: google.protobuf.Struct configuration_patch = 7;
   */
  configurationPatch?: JsonObject | undefined;

  /**
   * @generated from field: repeated string clear_configuration_fields = 8;
   */
  clearConfigurationFields: string[];
};

/**
 * Describes the message nama.api.v1.UpdateProviderInstanceRequest.
 * Use `create(UpdateProviderInstanceRequestSchema)` to create a new message.
 */
export declare const UpdateProviderInstanceRequestSchema: GenMessage<UpdateProviderInstanceRequest>;

/**
 * @generated from message nama.api.v1.UpdateProviderInstanceResponse
 */
export declare type UpdateProviderInstanceResponse = Message<"nama.api.v1.UpdateProviderInstanceResponse"> & {
  /**
   * @generated from field: nama.api.v1.ProviderInstance provider_instance = 1;
   */
  providerInstance?: ProviderInstance | undefined;
};

/**
 * Describes the message nama.api.v1.UpdateProviderInstanceResponse.
 * Use `create(UpdateProviderInstanceResponseSchema)` to create a new message.
 */
export declare const UpdateProviderInstanceResponseSchema: GenMessage<UpdateProviderInstanceResponse>;

/**
 * @generated from message nama.api.v1.DeleteProviderInstanceRequest
 */
export declare type DeleteProviderInstanceRequest = Message<"nama.api.v1.DeleteProviderInstanceRequest"> & {
  /**
   * @generated from field: string operation_id = 1;
   */
  operationId: string;

  /**
   * @generated from field: string provider_instance_id = 2;
   */
  providerInstanceId: string;

  /**
   * @generated from field: string expected_revision = 3;
   */
  expectedRevision: string;
};

/**
 * Describes the message nama.api.v1.DeleteProviderInstanceRequest.
 * Use `create(DeleteProviderInstanceRequestSchema)` to create a new message.
 */
export declare const DeleteProviderInstanceRequestSchema: GenMessage<DeleteProviderInstanceRequest>;

/**
 * @generated from message nama.api.v1.DeleteProviderInstanceResponse
 */
export declare type DeleteProviderInstanceResponse = Message<"nama.api.v1.DeleteProviderInstanceResponse"> & {
};

/**
 * Describes the message nama.api.v1.DeleteProviderInstanceResponse.
 * Use `create(DeleteProviderInstanceResponseSchema)` to create a new message.
 */
export declare const DeleteProviderInstanceResponseSchema: GenMessage<DeleteProviderInstanceResponse>;

/**
 * @generated from enum nama.api.v1.ProviderCapability
 */
export enum ProviderCapability {
  /**
   * @generated from enum value: PROVIDER_CAPABILITY_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: PROVIDER_CAPABILITY_LIBRARY_READ = 1;
   */
  LIBRARY_READ = 1,

  /**
   * @generated from enum value: PROVIDER_CAPABILITY_ARTWORK_RESOLVE = 2;
   */
  ARTWORK_RESOLVE = 2,

  /**
   * @generated from enum value: PROVIDER_CAPABILITY_PLAYBACK_PLAN = 3;
   */
  PLAYBACK_PLAN = 3,

  /**
   * @generated from enum value: PROVIDER_CAPABILITY_PLAYBACK_OPEN = 4;
   */
  PLAYBACK_OPEN = 4,

  /**
   * @generated from enum value: PROVIDER_CAPABILITY_PLAYBACK_REPORT = 5;
   */
  PLAYBACK_REPORT = 5,

  /**
   * @generated from enum value: PROVIDER_CAPABILITY_PLAYBACK_REPORTS_USER_STATE = 6;
   */
  PLAYBACK_REPORTS_USER_STATE = 6,

  /**
   * @generated from enum value: PROVIDER_CAPABILITY_WATCH_STATE_READ = 7;
   */
  WATCH_STATE_READ = 7,

  /**
   * @generated from enum value: PROVIDER_CAPABILITY_WATCHED_WRITE = 8;
   */
  WATCHED_WRITE = 8,

  /**
   * @generated from enum value: PROVIDER_CAPABILITY_PROGRESS_WRITE = 9;
   */
  PROGRESS_WRITE = 9,
}
// ...
export enum ProviderConnectionStatus {
  /**
   * @generated from enum value: PROVIDER_CONNECTION_STATUS_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: PROVIDER_CONNECTION_STATUS_CONNECTED = 1;
   */
  CONNECTED = 1,

  /**
   * @generated from enum value: PROVIDER_CONNECTION_STATUS_AUTHENTICATION_FAILED = 2;
   */
  AUTHENTICATION_FAILED = 2,

  /**
   * @generated from enum value: PROVIDER_CONNECTION_STATUS_UNREACHABLE = 3;
   */
  UNREACHABLE = 3,

  /**
   * @generated from enum value: PROVIDER_CONNECTION_STATUS_INCOMPATIBLE = 4;
   */
  INCOMPATIBLE = 4,
}

/**
 * Describes the enum nama.api.v1.ProviderConnectionStatus.
 */
export declare const ProviderConnectionStatusSchema: GenEnum<ProviderConnectionStatus>;

/**
 * @generated from service nama.api.v1.ProviderService
 */
export declare const ProviderService: GenService<{
  /**
   * @generated from rpc nama.api.v1.ProviderService.ListProviderTypes
   */
  listProviderTypes: {
    methodKind: "unary";
    input: typeof ListProviderTypesRequestSchema;
    output: typeof ListProviderTypesResponseSchema;
  },
  /**
   * @generated from rpc nama.api.v1.ProviderService.ListProviderInstances
   */
  listProviderInstances: {
    methodKind: "unary";
    input: typeof ListProviderInstancesRequestSchema;
    output: typeof ListProviderInstancesResponseSchema;
  },
  /**
   * @generated from rpc nama.api.v1.ProviderService.GetProviderInstance
   */
  getProviderInstance: {
    methodKind: "unary";
    input: typeof GetProviderInstanceRequestSchema;
    output: typeof GetProviderInstanceResponseSchema;
  },
  /**
   * @generated from rpc nama.api.v1.ProviderService.TestProviderConfiguration
   */
  testProviderConfiguration: {
    methodKind: "unary";
    input: typeof TestProviderConfigurationRequestSchema;
    output: typeof TestProviderConfigurationResponseSchema;
  },
  /**
   * @generated from rpc nama.api.v1.ProviderService.CreateProviderInstance
   */
  createProviderInstance: {
    methodKind: "unary";
    input: typeof CreateProviderInstanceRequestSchema;
    output: typeof CreateProviderInstanceResponseSchema;
  },
  /**
   * @generated from rpc nama.api.v1.ProviderService.UpdateProviderInstance
   */
  updateProviderInstance: {
    methodKind: "unary";
    input: typeof UpdateProviderInstanceRequestSchema;
    output: typeof UpdateProviderInstanceResponseSchema;
  },
  /**
   * @generated from rpc nama.api.v1.ProviderService.TestProviderInstance
   */
  testProviderInstance: {
    methodKind: "unary";
    input: typeof TestProviderInstanceRequestSchema;
    output: typeof TestProviderInstanceResponseSchema;
  },
  /**
   * @generated from rpc nama.api.v1.ProviderService.DeleteProviderInstance
   */
  deleteProviderInstance: {
    methodKind: "unary";
    input: typeof DeleteProviderInstanceRequestSchema;
    output: typeof DeleteProviderInstanceResponseSchema;
  },
}>;